pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ads-scanner"
version = "0.1.0"
description = "Attribution Density Scanner (ADS)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "ADS Contributors" }]
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.111.0",
  "joblib>=1.4.2",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "scikit-learn>=1.4.0",
  "typer>=0.12.3",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.4.0",
  "isort>=5.13.2",
  "mypy>=1.10.0",
  "pre-commit>=3.7.1",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "types-PyYAML>=6.0.12.20250915",
]
trak = ["torch>=2.2.0"]
cea = []
dda = []

[project.scripts]
ads = "ads.cli:app"

[tool.setuptools]
packages = ["ads"]

[tool.setuptools.package-data]
ads = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "D"]
ignore = ["D203", "D213", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["D103"]
"tests/*.py" = ["D103"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
